Refector: 댓글 ReplyInput 코드 분리
